move reports logic to seperate service
